requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
Pillow>=8.4.0
tqdm>=4.62.0
tensorboard>=2.7.0
scikit-image>=0.19.0
matplotlib>=3.5.0
timm>=0.4.12
torchinfo>=1.5.4